requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
ecdsa==0.19.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
setuptools==75.8.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
ulid-py==1.1.0
unicorn==2.1.2
uvicorn==0.34.0
wheel==0.45.1
